requirements.txt
snowflake-connector-python
streamlit
replicate
python-decouple
openai